+++ OptimoveSDK/optimove-sdk/src/main/java/com/optimove/android/optimobile/Optimobile.java
@@ -258,6 +258,10 @@ public final class Optimobile {
 
         if (isNewUserIdentifier) {
             OptimoveInApp.getInstance().handleInAppUserChange(context, Optimove.getConfig());
+
+            if (Optimove.getConfig().isOverlayMessagingEnabled()) {
+                OptimoveOverlayMessaging.getInstance().resetSession();
+            }
         }
     }